customer/src/main/java/com/vipulbehl/customer/CustomerService.java
package com.vipulbehl.customer;

import org.springframework.stereotype.Service;

import java.util.List;

@Service
public record CustomerService(CustomerRepository customerRepository) {
    public Customer registerCustomer(CustomerRegistrationRequest customerRegistrationRequest) {
        Customer customer = Customer.builder()
                .firstName(customerRegistrationRequest.firstName())
                .lastName(customerRegistrationRequest.lastName())
                .email(customerRegistrationRequest.email())
                .build();

        customerRepository.save(customer);
        return customer;
    }

    public List<Customer> listCustomers() {
        return customerRepository.findAll();
    }

    public Customer getCustomerById(int id) {
        return customerRepository.getById(id);
    }

    public void deleteCustomerById(Integer id) {
        customerRepository.deleteById(id);
    }
}
